Renamed contacts module to address_book
